+++ Faker/Business.cs
@@ -1,23 +1,14 @@
-﻿using RimuTec.Faker.Extensions;
-using RimuTec.Faker.Helper;
+﻿using RimuTec.Faker.Helper;
 using System;
-using YamlDotNet.Serialization;
 
 namespace RimuTec.Faker
 {
    /// <summary>
    /// Generators for business related data
    /// </summary>
-   public static class Business
+   public class Business : GeneratorBase
    {
-      // Resources used by this class from https://github.com/stympy/faker/blob/master/lib/locales/en/business.yml
-
-      static Business()
-      {
-         const string yamlFileName = "RimuTec.Faker.locales.en.business.yml";
-         locale locale = YamlLoader.Read<locale>(yamlFileName);
-         _business = locale.en.faker.business;
-      }
+      private Business() { }
 
       /// <summary>
       /// Generates a credit card expiry date.
@@ -34,7 +25,7 @@ namespace RimuTec.Faker
       /// <returns></returns>
       public static string CreditCardNumber()
       {
-         return _business.CreditCardNumbers.Sample();
+         return YamlLoader.Fetch("business.credit_card_numbers");
       }
 
       /// <summary>
@@ -43,38 +34,7 @@ namespace RimuTec.Faker
       /// <returns></returns>
       public static string CreditCardType()
       {
-         return _business.CreditCardTypes.Sample();
+         return YamlLoader.Fetch("business.credit_card_types");
       }
-
-      private static readonly business _business;
-
-#pragma warning disable IDE1006 // Naming Styles
-      // Helper classes for reading the yaml file. Note that the class names are
-      // intentionally lower case.
-
-      internal class locale
-      {
-         public en en { get; set; }
-      }
-
-      internal class en
-      {
-         public faker faker { get; set; }
-      }
-
-      internal class faker
-      {
-         public business business { get; set; }
-      }
-
-      internal class business
-      {
-         [YamlMember(Alias = "credit_card_numbers", ApplyNamingConventions = false)]
-         public string[] CreditCardNumbers { get; set; }
-
-         [YamlMember(Alias = "credit_card_types", ApplyNamingConventions = false)]
-         public string[] CreditCardTypes { get; set; }
-      }
-#pragma warning restore IDE1006 // Naming Styles
    }
 }